+++ Common/ItemCommon/VariantItemRenderer.cs
@@ -10,13 +10,18 @@ public sealed class VariantItemRenderer : GlobalItem
 	public static readonly int[] VariantCounts = ItemID.Sets.Factory.CreateIntSet();
 
 	#region helpers
-	public static int GetVariant(Item item) => (item.TryGetGlobalItem(out VariantItemRenderer global) ? global.subID : 0) % VariantCounts[item.type];
+	public static int GetVariant(Item item) => (item.TryGetGlobalItem(out VariantItemRenderer global) ? global.subID ?? 0 : 0) % VariantCounts[item.type];
 
 	public static Texture2D GetTexture(Item item, out Rectangle source)
 	{
 		if (item.ModItem is ModItem modItem && VariantCounts[item.type] > 0)
 		{
-			Texture2D result = AssetLoader.GetTexture(modItem.Name, modItem.Texture + GetVariant(item)).Value;
+			string texturePath = modItem.Texture;
+
+			if (texturePath[texturePath.Length - 1] == '0') //If the last character is zero (one of the variants), remove the character
+				texturePath = texturePath.Remove(texturePath.Length - 1);
+
+			Texture2D result = AssetLoader.GetTexture(modItem.Name + GetVariant(item), texturePath + GetVariant(item)).Value;
 			source = (Main.itemAnimations[item.type] != null) ? Main.itemAnimations[item.type].GetFrame(result) : result.Frame();
 
 			return result;
@@ -31,26 +36,28 @@ public sealed class VariantItemRenderer : GlobalItem
 
 	public override bool InstancePerEntity => true;
 
-	public int subID = -1;
+	public int? subID;
 
 	public override bool AppliesToEntity(Item entity, bool lateInstantiation) => VariantCounts[entity.type] > 0;
 
+	public override void OnStack(Item destination, Item source, int numToTransfer)
+	{
+		if (source.TryGetGlobalItem(out VariantItemRenderer global))
+			subID = global.subID;
+	}
+
 	public override bool PreDrawInWorld(Item item, SpriteBatch spriteBatch, Color lightColor, Color alphaColor, ref float rotation, ref float scale, int whoAmI)
 	{
-		if (subID == -1)
-			subID = Main.rand.Next(VariantCounts[item.type]);
+		subID ??= Main.rand.Next(VariantCounts[item.type]);
 
-		Texture2D texture = GetTexture(item, out Rectangle source);
-		Vector2 position = item.position + item.Size - source.Size() - Main.screenPosition;
-
-		spriteBatch.Draw(texture, position, source, GetAlpha(item, lightColor) ?? lightColor, rotation, Vector2.Zero, scale, 0, 0);
+		Texture2D texture = GetTexture(item, out _);
+		ItemMethods.DrawInWorld(item, lightColor, rotation, scale, texture);
 		return false;
 	}
 
 	public override bool PreDrawInInventory(Item item, SpriteBatch spriteBatch, Vector2 position, Rectangle frame, Color drawColor, Color itemColor, Vector2 origin, float scale)
 	{
-		if (subID == -1)
-			subID = Main.rand.Next(VariantCounts[item.type]);
+		subID ??= Main.rand.Next(VariantCounts[item.type]);
 
 		Texture2D texture = GetTexture(item, out Rectangle source);
