+++ Content/Desert/ScarabBoss/Items/Projectiles/AdornedBowHeld.cs
@@ -2,6 +2,7 @@
 using SpiritReforged.Common.PrimitiveRendering;
 using SpiritReforged.Common.PrimitiveRendering.PrimitiveShape;
 using SpiritReforged.Common.ProjectileCommon.Abstract;
+using SpiritReforged.Common.Visuals;
 using SpiritReforged.Common.Visuals.Glowmasks;
 using Terraria;
 using Terraria.Audio;
@@ -76,8 +77,8 @@ public class AdornedBowHeld() : ChargeBowProjectile(1.15f, 2f, 40)
 		{
 			_flashTimer--;
 
-			Lighting.AddLight(Projectile.Center, 
-				AdornedBowGlobalProjectile.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, [Color.Magenta, Color.Orange, Color.Cyan]).ToVector3()
+			Lighting.AddLight(Projectile.Center,
+				DrawHelpers.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, [Color.Magenta, Color.Orange, Color.Cyan]).ToVector3()
 				* 0.5f * (_flashTimer / (float)MAX_FLASH_TIMER));
 		}
 		
@@ -112,7 +113,7 @@ public class AdornedBowHeld() : ChargeBowProjectile(1.15f, 2f, 40)
 		Texture2D glowmaskTex = glowmaskInfo.Glowmask.Value;
 		Texture2D starTex = AssetLoader.LoadedTextures["StarChromatic"].Value;
 
-		Color color = (_flashTimer > 0f ? Color.Lerp(AdornedBowGlobalProjectile.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, _palette.Colors), Color.LightSteelBlue, 1f - _flashTimer / (float)MAX_FLASH_TIMER) : Color.LightSteelBlue).Additive();
+		Color color = (_flashTimer > 0f ? Color.Lerp(DrawHelpers.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, _palette.Colors), Color.LightSteelBlue, 1f - _flashTimer / (float)MAX_FLASH_TIMER) : Color.LightSteelBlue).Additive();
 		float perfectShotProgress = EaseSine.Ease(EaseCircularOut.Ease(1 - _perfectShotCurTimer / _perfectShotMaxTime));
 		float strength = Charge * (Projectile.timeLeft / 30f);
 
@@ -177,7 +178,7 @@ public class AdornedBowHeld() : ChargeBowProjectile(1.15f, 2f, 40)
 
 		var square = new SquarePrimitive
 		{
-			Color =  (_flashTimer > 0f ? Color.Lerp(AdornedBowGlobalProjectile.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, _palette.Colors), Color.LightSteelBlue, 1f - _flashTimer / (float)MAX_FLASH_TIMER) : Color.LightSteelBlue).Additive() * Projectile.Opacity,
+			Color =  (_flashTimer > 0f ? Color.Lerp(DrawHelpers.MulticolorLerp(_flashTimer / (float)MAX_FLASH_TIMER, _palette.Colors), Color.LightSteelBlue, 1f - _flashTimer / (float)MAX_FLASH_TIMER) : Color.LightSteelBlue).Additive() * Projectile.Opacity,
 			Height = dimensions.X,
 			Length = dimensions.Y,
 			Position = Projectile.Center + new Vector2(0f, Projectile.gfxOffY) - Main.screenPosition + Vector2.UnitX.RotatedBy(Projectile.rotation) * 5,
