+++ Content/Forest/ButterflyStaff/ButterflyCritter.cs
@@ -1,6 +1,6 @@
 using SpiritReforged.Common.NPCCommon;
 using SpiritReforged.Common.WorldGeneration.Microbiomes;
-using SpiritReforged.Common.WorldGeneration.Microbiomes.Biomes;
+using SpiritReforged.Common.WorldGeneration.Micropasses.Passes;
 using Terraria.Audio;
 using Terraria.GameContent.Bestiary;
 
@@ -203,7 +203,7 @@ public class ButterflyCritter : ModNPC
 			var coord = new Point(spawnInfo.SpawnTileX, spawnInfo.SpawnTileY);
 			foreach (var biome in MicrobiomeSystem.Microbiomes)
 			{
-				if (biome is not ButterflyShrineBiome e)
+				if (biome is not ButterflyMicropass.ButterflyShrineBiome e)
 					continue;
 
 				if (e.Rectangle.Contains(coord))
