+++ Content/Forest/ButterflyStaff/ButterflyStump.cs
@@ -1,9 +1,7 @@
 using RubbleAutoloader;
 using SpiritReforged.Common.ItemCommon;
-using SpiritReforged.Common.TileCommon;
-using SpiritReforged.Common.Visuals.Glowmasks;
 using SpiritReforged.Common.WorldGeneration.Microbiomes;
-using SpiritReforged.Common.WorldGeneration.Microbiomes.Biomes;
+using SpiritReforged.Common.WorldGeneration.Micropasses.Passes;
 using System.Linq;
 using Terraria.DataStructures;
 using Terraria.GameContent.ObjectInteractions;
@@ -54,7 +52,7 @@ public class ButterflyStump : ModTile, IAutoloadRubble
 
 	public override void KillMultiTile(int i, int j, int frameX, int frameY)
 	{
-		var overlap = MicrobiomeSystem.Microbiomes.Where(x => x is ButterflyShrineBiome e && e.Rectangle.Contains(new Point(i, j)));
+		var overlap = MicrobiomeSystem.Microbiomes.Where(x => x is ButterflyMicropass.ButterflyShrineBiome e && e.Rectangle.Contains(new Point(i, j)));
 		bool removedAny = false;
 
 		foreach (var biome in overlap)
