+++ Content/Ziggurat/Biome/ZigguratGlobalNPC.cs
@@ -1,4 +1,4 @@
-﻿using SpiritReforged.Common.WorldGeneration.Microbiomes.Biomes.Ziggurat;
+﻿using SpiritReforged.Common.WorldGeneration.Micropasses.Passes.Ziggurat;
 using System.Linq;
 
 namespace SpiritReforged.Content.Ziggurat.Biome;
@@ -10,7 +10,7 @@ internal class ZigguratGlobalNPC : GlobalNPC
 		Point tileCoords = new(spawnInfo.SpawnTileX, spawnInfo.SpawnTileY - 1);
 		int wallType = Framing.GetTileSafely(tileCoords).WallType;
 
-		return !Main.wallHouse[wallType] && ZigguratMicrobiome.TotalBounds.Any(x => x.Contains(tileCoords));
+		return !Main.wallHouse[wallType] && ZigguratMicropass.TotalBounds.Any(x => x.Contains(tileCoords));
 	}
 
 	public override void EditSpawnPool(IDictionary<int, float> pool, NPCSpawnInfo spawnInfo)
