+++ Localization/pt-BR/Mods.SpiritReforged.GenConfigs.hjson
@@ -639,6 +639,71 @@ Pages: {
 				DisplayName: "[i:SpiritReforged/LapisDresserItem] Quantidade de Tesouros em Cômodas"
 				Tooltip: A faixa de variação do número de itens em uma cômoda.
 			}
+
+			DebrisMinimum: {
+				// DisplayName: DebrisMinimum
+				// Tooltip: DebrisMinimum
+			}
+
+			DebrisRange: {
+				// DisplayName: DebrisRange
+				// Tooltip: DebrisRange
+			}
+
+			TopPlatChance: {
+				// DisplayName: TopPlatChance
+				// Tooltip: TopPlatChance
+			}
+
+			BottomPlatChance: {
+				// DisplayName: BottomPlatChance
+				// Tooltip: BottomPlatChance
+			}
+
+			FlagpoleChance: {
+				// DisplayName: FlagpoleChance
+				// Tooltip: FlagpoleChance
+			}
+
+			FlagpoleCountMinumum: {
+				// DisplayName: FlagpoleCountMinumum
+				// Tooltip: FlagpoleCountMinumum
+			}
+
+			FlagpoleCountRange: {
+				// DisplayName: FlagpoleCountRange
+				// Tooltip: FlagpoleCountRange
+			}
+
+			FlagpoleHeightMin: {
+				// DisplayName: FlagpoleHeightMin
+				// Tooltip: FlagpoleHeightMin
+			}
+
+			FlagpoleHeightRange: {
+				// DisplayName: FlagpoleHeightRange
+				// Tooltip: FlagpoleHeightRange
+			}
+
+			PotMultiplier: {
+				// DisplayName: PotMultiplier
+				// Tooltip: PotMultiplier
+			}
+
+			ChestChance: {
+				// DisplayName: ChestChance
+				// Tooltip: ChestChance
+			}
+
+			UncommonPotChance: {
+				// DisplayName: UncommonPotChance
+				// Tooltip: UncommonPotChance
+			}
+
+			AbovegroundPillarChance: {
+				// DisplayName: AbovegroundPillarChance
+				// Tooltip: AbovegroundPillarChance
+			}
 		}
 
 		Presets: {
