+++ Localization/zh-Hans/Mods.SpiritReforged.GenConfigs.hjson
@@ -639,6 +639,71 @@ Pages: {
 				DisplayName: "[i:SpiritReforged/LapisDresserItem] 梳妆台战利品数量"
 				Tooltip: 梳妆台中物品数量的范围值。
 			}
+
+			DebrisMinimum: {
+				// DisplayName: DebrisMinimum
+				// Tooltip: DebrisMinimum
+			}
+
+			DebrisRange: {
+				// DisplayName: DebrisRange
+				// Tooltip: DebrisRange
+			}
+
+			TopPlatChance: {
+				// DisplayName: TopPlatChance
+				// Tooltip: TopPlatChance
+			}
+
+			BottomPlatChance: {
+				// DisplayName: BottomPlatChance
+				// Tooltip: BottomPlatChance
+			}
+
+			FlagpoleChance: {
+				// DisplayName: FlagpoleChance
+				// Tooltip: FlagpoleChance
+			}
+
+			FlagpoleCountMinumum: {
+				// DisplayName: FlagpoleCountMinumum
+				// Tooltip: FlagpoleCountMinumum
+			}
+
+			FlagpoleCountRange: {
+				// DisplayName: FlagpoleCountRange
+				// Tooltip: FlagpoleCountRange
+			}
+
+			FlagpoleHeightMin: {
+				// DisplayName: FlagpoleHeightMin
+				// Tooltip: FlagpoleHeightMin
+			}
+
+			FlagpoleHeightRange: {
+				// DisplayName: FlagpoleHeightRange
+				// Tooltip: FlagpoleHeightRange
+			}
+
+			PotMultiplier: {
+				// DisplayName: PotMultiplier
+				// Tooltip: PotMultiplier
+			}
+
+			ChestChance: {
+				// DisplayName: ChestChance
+				// Tooltip: ChestChance
+			}
+
+			UncommonPotChance: {
+				// DisplayName: UncommonPotChance
+				// Tooltip: UncommonPotChance
+			}
+
+			AbovegroundPillarChance: {
+				// DisplayName: AbovegroundPillarChance
+				// Tooltip: AbovegroundPillarChance
+			}
 		}
 
 		Presets: {
